+++ frontend/src/components/dialog/dataType/AddDataType.tsx
@@ -2,10 +2,9 @@ import React, {useState} from 'react';
 import {ProForm, ModalForm, ProFormText} from '@ant-design/pro-components';
 import useProjectStore from "@/store/project/useProjectStore";
 import shallow from "zustand/shallow";
-import {Divider} from '@mui/material';
 import _ from "lodash";
 import {PlusOutlined} from "@ant-design/icons";
-import {Button} from "antd";
+import {Button, Divider} from "antd";
 
 export type AddDataTypeProps = {
   moduleDisable: boolean;